requirements.txt
absl-py==0.1.6
affine==2.1.0
appnope==0.1.0
astor==0.6.2
astroid==1.6.1
backports.weakref==1.0.post1
bleach==1.5.0
boto==2.48.0
boto3==1.5.24
botocore==1.8.38
bz2file==0.98
certifi==2017.11.5
chardet==3.0.4
click==6.7
click-plugins==1.0.3
cligj==0.4.0
config==0.3.9
coverage==4.4.2
cycler==0.10.0
decorator==4.1.2
docutils==0.14
entrypoints==0.2.3
enum34==1.1.6
gast==0.2.0
gensim==3.3.0
gmpy2==2.0.8
graphviz==0.8.2
grpcio==1.10.0
h5py==2.7.1
html5lib==0.9999999
idna==2.6
ipykernel==4.8.2
ipython==6.2.1
ipython-genutils==0.2.0
ipywidgets==7.1.2
isort==4.3.4
jedi==0.11.1
jieba==0.39
Jinja2==2.10
jmespath==0.9.3
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.2
jupyter-console==5.2.0
jupyter-core==4.4.0
Keras==2.1.5
kiwisolver==1.0.1
lazy-object-proxy==1.3.1
Markdown==2.6.11
MarkupSafe==1.0
matplotlib==2.1.2
matplotlylib==0.1.0
mccabe==0.6.1
mistune==0.8.3
mock==2.0.0
nbconvert==5.3.1
nbformat==4.4.0
notebook==5.4.0
numpy==1.14.2
olefile==0.44
opencv-python==3.4.0.12
pandas==0.22.0
pandocfilters==1.4.2
parso==0.1.1
pbr==3.1.1
pexpect==4.4.0
pickleshare==0.7.4
Pillow==5.0.0
plotly==2.2.2
prompt-toolkit==1.0.15
protobuf==3.5.0.post1
ptyprocess==0.5.2
pydot==1.2.4
Pygments==2.2.0
pylint==1.8.2
pyparsing==2.2.0
python-dateutil==2.6.1
pytz==2017.3
PyYAML==3.12
pyzmq==17.0.0
qtconsole==4.3.1
rasterio==0.36.0
requests==2.18.4
s3transfer==0.1.12
scikit-learn==0.19.1
scipy==1.0.0
Send2Trash==1.5.0
simplegeneric==0.8.1
six==1.11.0
sklearn==0.0
smart-open==1.5.6
snownlp==0.12.3
snuggs==1.4.1
tensorboard==1.6.0
tensorflow==1.7.0rc1
tensorflow-tensorboard==1.5.1
termcolor==1.1.0
terminado==0.8.1
testpath==0.3.1
tflearn==0.3.2
thulac==0.1.2
tornado==5.0
traitlets==4.3.2
unrar==0.3
urllib3==1.22
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.13
widgetsnbextension==3.1.4
wrapt==1.10.11
